party speed to float

+++ src/Application/Game.cpp
@@ -1112,7 +1112,7 @@ void Game::processQueuedMessages() {
                 SaveGame(1, 0);
 
                 pParty->pos = Vec3i(-17331, 12547, 465); // respawn point in Harmondale
-                pParty->speed = Vec3i();
+                pParty->speed = Vec3f();
                 pParty->_viewYaw = 0;
                 pParty->uFallStartZ = pParty->pos.z;
                 pParty->_viewPitch = 0;
@@ -2105,7 +2105,7 @@ void Game::gameLoop() {
                 Source = pLocationName;
                 pParty->uFallStartZ = pParty->pos.z;
                 pParty->_viewPitch = 0;
-                pParty->speed = Vec3i();
+                pParty->speed = Vec3f();
                 // change map
                 if (pCurrentMapName != Source) {
                     pCurrentMapName = Source;
